SemanticPass: Call sa_MemberExpression in sa_CallExpression
When callee of a CallExpression node is a MemberExpression node, process
it using sa_MemberExpression first (via sa_expr) and then investigate
whether it represents a reducer/function or not. This means we lean on
sa_MemberExpression processing instead of doing our own in
sa_CallExpression. This makes sa_CallExpression code more generic and
the language more regular.

A nice side effect of the change is fixing an incorrect error message.