compiler: Support expression statements
Expression statements replace more limited assignment statements.

Note the change required a small reordering in the FunctionStatement
rule in the grammar. This was necessary because ExpressionStatement
matches things like "return;" which should be parsed by other *Statement
rules. The correct parsing is ensured by putting ExpressionStatement as
the last choice in FunctionStatement.