join.js - add interval info to marks
Join accepts batched streams, but only uses marks from input to set the
time of points in groups to mark.time+-epsilon and to flag the input
as batched. The marks themselves are dropped before joining.

Only after the joined points are emitted, if the appropriate inputs were
batched (see batched_leader() for the logic), it restores the
surrounding marks again. In this case, the interval boundaries
will be start_time and complete_time of the joined group. Marks
will contain the interval info.