SemanticPass: Process argument of the "*" operator in correct mode
Consider an expression like this:

    *field = "foo"

The "field" variable is evaluated and its value is used to select a
point field. It is not itself assigned to. In semantic pass, this
corresponds to processing with this.lhs == false.

Before this commit, code in sa_UnaryExpression wrongly processed "*"
arguments with the same this.lhs value as the operator itself. This
commit fixes it to always process them with this.lhs == false.