SemanticPass: Introduce new "update" expression mode
Introduce new "update" expression mode and use it to process arguments
of the "++" and "--" operators (both prefix and postfix), which were
previously processed in the "assign" mode. The change reflects the fact
that arguments of these operators are slightly different than assignment
LHS (unlike it, they are both read from and written to). It will also
help with future refactoring.