Build, GraphCompiler: Change how assignments are compiled
Before this commit, gen_AssignmentExpression just compiled code for both
sides of the assignment and joined the results using the "=" operator.
This worked great but it created a limitation for implementing type-safe
member assignment. To make it work, expressions like:

    a[1] = 5

will need to be compiled into something like this:

    juttle.ops.set(a_1, 1, 5)

The actual JavaScript-level assignment would not be exposed in compiled
code but hidden inside juttle.ops.set.

To make it possible to compile assignments this way the whole model has
to change, which this commit does. From now on, gen_AssignmentExpression
compiles just the RHS and then sends resulting code to the method
compiling the LHS together with the "lhs" flag. It is up to this method
to compile the overall assignment and embed generated RHS code into the
result.

The principial advantage of this model is that gen_AssignmentExpression
doesn't have to know how to implement assignment for various possible
LHSs (variables, fields, member expressions,...) -- this knowledge is
distributed to methods responsible for individual constructs. The result
is a flexible and extensible design.