Misc updates to support new file import behavior.
With the new local import rules, relative paths must begin with ./. So
fix any paths implied to be relative imports to use ./.

In the cli's show_in_context function, don't assume that the main
program has a filename of 'main'. Instead check against the set of
saved module filenames and consider it a module if found, the main
program otherwise.

Now that we have local imports, we don't want to use cwd for system
imports, so remove it from the set of search paths.

Also get examples-check working:
 - add the filename to all .juttle files so file-relative imports work.
 - module files can't be compiled on their own. So add a convention
   that files containing the string 'examples-check:ignore' will be
   skipped.