Update resolver interface to prepare for file-local imports.
Modify the resolver api to pass along an importer path along with the
module path and module name. For file imports, this should be the path
to the file doing the import.

The thing returned by the resolver api is now an object with source:
juttle source code + name: the name of the thing chosen while
resolving. For URLs, name is the url. For files, the file is the full
path to the filename that was found.

The parser takes advantage of the returned module name by using it as
the filename when parsing the module source code. That way, any
imports will have the importer path set correctly.

This doesn't actually make any changes to the file resolver's
behavior. That will be in a later commit.

Also update bin/juttle to pass the filename as the importer_path--the
compiler already supported an option 'filename' which was passed along
to the parser. In bin/juttle, specify the filename (fully resolved
using path.resolve()) for any of the modes where the source of the
juttle program is a file.