Use instanceOf JuttleError to identify Juttle Errors.
Instead of assuming that any error with a location is a Juttle Error,
use instanceof JuttleError instead.

The work of dealing with the error is in a common function
handle_juttle_error() in lib/cli/errors. This is thinking ahead to
calls show_in_context when the error has a location, and just
generically prints the error otherwise. Both versions rely on a common
(not-exported) function to format the error string.