SemanticPass: Move mutability check in sa_Variable
Move mutability check in sa_Variable to a place where variable's symbol
is available. This will allow to rewrite it as a symbol check.

The move is correct because if the original position of the check was
reached, then sombol lookup would succeed and the new position would be
also reached. In other words, the check is invoked under exactly the
same circumstances in both positions.