compiler: Make the "[]" operator work properly in lvalues
This commit adds tests that specify behavior of the "[]" operator in
lvalues and adjusts the implementation to match. Specifically, it adds
missing type checks into ops.set and allows modification of expressions
not associated with a symbol (which enables nested member assignment).

Note that member assignment in "put"/"reduce" still isn't allowed.
Making it work requires more substantial changes which will be done
later.