Add support for file local resolving in the file resolver.
In the file resolver, break _search into _search_local_import and
_search_system_import. When module_path begins with /, ./, or ../,
it's a local import and the file resolver calls
_search_local_import. Otherwise, it's a system import and the file
resolver calls _search_system_import.

_canonicalize is used by both _search methods and does some common
work for the local and system resolvers, adding a .juttle suffix to
filenames if needed.

_search_local_import either takes the current working directory or the
dirname of importer_path and uses that to resolve module_path into a
full filename.

_search_system_import is unchanged, using the set of search paths and
trying to resolve the path suffix relative to each.

Only _search_local_import returns a name that's the fully resolved
path of the file. _search_system_import just returns the original
module path.