modules: support importing procs from native js modules
This adds rudimentary support for extending the runtime with procs
implemented in javascript by supporting `import "/path/to/file.js"`
in which the file exports an object containing the proc implementation.

To make this work, the module loader looks at the file extension, and for
.js files it simply inlines the source and returns it.

Then the semantic pass evals the blob of code and puts all the symbols from
the exported `procs` into scope.

The flowgraph builder generates a new type of natice proc call node and also
passes through the native modules so they can be carried through to the
built flowgraph.

Finally, the module source is inlined into the compiled code so it gets
evaluated again as part of the program itself.