Add support for adapter modules.
Create the convention that adapters can publish modules below the
directory 'juttle', which is a subdirectory below the directory
where the adapter was loaded (e.g. juttle-XXX-adapter/juttle).

In import statements, the file is imported using "import
adapters/XXX/path-to-module-file". Note that juttle is not directly
visible, replaced by the name of the adapter.

The file resolver calls adapters.list and considers any adapters that
are loaded and not built-in. For each adapter, it checks to see if
<adapter install path>/juttle exists as a directory and if so
adds it to a set of adapter module paths.

In _search_system_imports, any module_path starting with 'adapters/'
is an adapter module import and is passed off to
_search_adapter_import.

In _search_adapter_import, drop the initial 'adapters' from
module_path, take the first subdirectory, and see if it matches the
name of an adapter. If so, swap out the adapter name and the adapter
module path, and try to load the file.