GraphCompiler: Improve $lhs param building
When building the value of the $lhs param for "put" and "reduce", the
graph compiler assumed that LHS of their assignments will always be
either a Field or UnaryExpression node representing the assigned field.
With nested field assignment, situation will get more complicated
because MemberExpression will also be allowed.

This commit changes $lhs building code to cope with MemberExpression
nodes and renames $lhs to $assigned_fields because the old name would be
misleading in presence of nested field assignments.