Juttle grammar: Allow nested point field assignment
Note that the AssignmentLHS and StrictAssignmentLHS rules are now almost
the same. The only difference is that in StrictAssignmentLHS, the "base"
of the assignment is always a field reference. In the future, we may
decide to merge the two rules and move the field reference check to the
semantic pass.