SemanticPass: Fix reducer args coercion
Before this commit, variables passed as reducer args were coerced into
fields unless there was a const or var in scope with the same name. All
other entities such as functions or reducers were ignored. This was
inconsistent with by-lists and sort-lists where coercion in such cases
lead to a CANNOT-USE-AS-VARIABLE error.

This commit makes coercion of reducer args work exactly the same as in
by-lists and sort-lists (by utilizing the same code path).