Repackage non-juttle errors as juttle errors.
In the read adapter if the promise chain is caught with a non-juttle
error, wrap it in a runtime error.

This can happen if the file adapter tries to read a file that doesn't
exist. In this case the error will simply be an error from trying to
open the file and not a Juttle error.